Explicit cast from pointer addresses to int for x64 builds. Issue #33

+++ DICOM.Native/OpenJPEG/cio.c
@@ -98,7 +98,7 @@ void OPJ_CALLCONV opj_cio_close(opj_cio_t *cio) {
  * Get position in byte stream.
  */
 int OPJ_CALLCONV cio_tell(opj_cio_t *cio) {
-	return cio->bp - cio->start;
+	return (int)(cio->bp - cio->start);
 }
 
 /*
@@ -116,7 +116,7 @@ void OPJ_CALLCONV cio_seek(opj_cio_t *cio, int pos) {
  */
 int cio_numbytesleft(opj_cio_t *cio) {
   assert((cio->end - cio->bp) >= 0);
-	return cio->end - cio->bp;
+	return (int)(cio->end - cio->bp);
 }
 
 /*
